Init redis on bot start, rather than on class init
There was some offset from initalizing redis on the class start which resulted in the message listening task not firing, this forces redis to init once the bot is ready to initalize & there should be an established running loop for asyncio. I'm not sure if the issue was with asyncio, but I'm guessing it's related to it.