Limit string generation to 5 ahead of time rather than after
If a group has a ton of binds, this will block until all the bindings have a string generated for them - resulting in a long wait time (for groups with even as little as 13 bindings) and ratelimits being abused. This changes it so that the first 5 binds will have their string generated at most.